social media platform/backend/requirements.txt
Django>=4.2,<5
gunicorn
djangorestframework
django-cors-headers
